Refactored Storage to now provide getSqliteDatabase()
