+++ packages/d1/src/index.ts
@@ -1,2 +1,3 @@
 export * from "./database";
 export * from "./plugin";
+export * from "./statement";
